requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
fastapi
googletrans==4.0.0rc1
geocoder==1.38.1
openai==0.10.5
pickle-mixin==1.0.2
pydantic==1.9.0
azure-servicebus==7.8.3
opencensus-ext-azure==1.1.9